Extended user area with area groupings. Changed stuff so that it is clearer that the type column in user area should only contain USERAREA
