+++ LIQUIBASE/changelog/db-changelog-master.xml
@@ -16,5 +16,6 @@
     <include file="v2.0.6\db-changelog-addTableAreaUpdate.xml" relativeToChangelogFile="true"/>
     <include file="v2.0.10\db-changelog-changeDBTypesOnSomeAreaEntitys.xml" relativeToChangelogFile="true"/>
     <include file="v2.0.14\db-changelog-changePostgisStorageOfLargeAreas.xml" relativeToChangelogFile="true"/>
+    <include file="v2.1.3\db-changelog-extendUserAreasWithAreaGroup.xml" relativeToChangelogFile="true"/>
 
 </databaseChangeLog>
