--- rest/src/main/java/eu/europa/ec/fisheries/uvms/spatial/rest/mapper/AreaLocationMapper.java
@@ -1,105 +1,0 @@
-/*
-Developed by the European Commission - Directorate General for Maritime Affairs and Fisheries @ European Union, 2015-2016.
-
-This file is part of the Integrated Fisheries Data Management (IFDM) Suite. The IFDM Suite is free software: you can redistribute it 
-and/or modify it under the terms of the GNU General Public License as published by the Free Software Foundation, either version 3 of 
-the License, or any later version. The IFDM Suite is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; 
-without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more 
-details. You should have received a copy of the GNU General Public License along with the IFDM Suite. If not, see <http://www.gnu.org/licenses/>.
-
- */
-
-
-package eu.europa.ec.fisheries.uvms.spatial.rest.mapper;
-
-import eu.europa.ec.fisheries.uvms.spatial.model.schemas.*;
-import eu.europa.ec.fisheries.uvms.spatial.rest.dto.AreaCoordinateType;
-import eu.europa.ec.fisheries.uvms.spatial.rest.dto.GeoCoordinateType;
-import eu.europa.ec.fisheries.uvms.spatial.rest.dto.LocationQueryDto;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-public abstract class AreaLocationMapper {
-
-	private static final String AREA_TYPE = "areaType";
-	private static final String LOCATION_TYPE = "locationType";
-	public static final String AREA_ID = "id";
-
-	private static final AreaLocationMapper INSTANCE = null;
-
-
-	public static AreaLocationMapper mapper() {
-		return INSTANCE;
-	}
-
-	public abstract AreaTypeEntry getAreaTypeEntry(AreaCoordinateType areaDto);
-
-	public abstract LocationTypeEntry getLocationTypeEntry(LocationQueryDto locationDto);
-	
-
-	public abstract List<AreaTypeEntry> getAreaTypeEntryList(List<AreaCoordinateType> areaDtoList);
-	
-
-	
-	public abstract Coordinate getCoordinateFromDto(GeoCoordinateType geoCoordinateType);
-
-	protected AreaType getAreaTypeEnum(String areaType) {
-		if ( areaType == null ) {
-			return null;
-		}
-		return Enum.valueOf( AreaType.class, areaType.toUpperCase());
-	}
-	
-	protected Map<String, Object> extractProperties(AreaDetails areaDetails) {
-		Map<String, Object> propertyMap = new HashMap<>();
-		for (AreaProperty property : areaDetails.getAreaProperties()) {
-			Object propertyValue;
-
-			if ("scopeSelection".equalsIgnoreCase(property.getPropertyName()) ) {
-			} else {
-				propertyValue = property.getPropertyValue();
-			}
-
-		}
-		if (!propertyMap.isEmpty()) {
-			propertyMap.put(AREA_TYPE, String.valueOf(areaDetails.getAreaType().getAreaType()).toUpperCase());
-			propertyMap.put(AREA_ID, areaDetails.getAreaType().getId());
-		}
-		return propertyMap;
-	}
-	
-	protected List<Map<String, Object>> extractProperties(List<AreaDetails> areaDetailsList) {
-		List<Map<String, Object>> allPropertyMap = new ArrayList<>();
-		for (AreaDetails areaDetails : areaDetailsList) {
-			Map<String, Object> propertyMap = extractProperties(areaDetails);
-			allPropertyMap.add(propertyMap);
-		}
-		return allPropertyMap;
-	}
-	
-	protected String extractType(AreaDetails areaDetails) {
-		return String.valueOf(areaDetails.getAreaType().getAreaType());
-	}
-	
-	protected String extractType(AreaCoordinateType areaDto) {
-		return areaDto.getAreaType();
-	}
-	
-	protected Map<String, Object> extractProperties(LocationDetails locationDetails) {
-		Map<String, Object> propertyMap = new HashMap<>();
-		for (LocationProperty property : locationDetails.getLocationProperties()) {
-			propertyMap.put(property.getPropertyName(), property.getPropertyValue());
-		}
-		if (!propertyMap.isEmpty()) {
-			propertyMap.put(LOCATION_TYPE, String.valueOf(locationDetails.getLocationType().getLocationType()).toUpperCase());
-		}
-		return propertyMap;
-	}
-
-	protected String extractType(LocationDetails locationDetails) {
-		return String.valueOf(locationDetails.getLocationType().getLocationType());
-	}
-}
